refactor(signal/scan.py): move SeisOutFile
Move SeisOutFile to io module and perform some renaming. This change is
entirely internal and tested, it should not affect end user usage.
